requirements.txt
Django==2.1.7
django-cors-headers==3.0.1
django-filter==2.1.0
djangorestframework==3.9.4
django-heroku==0.3.1
Pillow==3.1.1
gunicorn==19.7
whitenoise==3.3.1
psycopg2==2.8.2
